+++ tests/tools/oecertdump/README.md
@@ -1,8 +1,8 @@
 oecertdump
 =====
 
-oecertdump is a utility that generates and validates report and certificates.
-Generation and validation logs are save to a log file specified by --out option.
+oecertdump is a utility that generates and validates reports and certificates.
+Generation and validation logs are saved to a log file specified by --out option.
 Default log filename is "oecertdump_out.log".
 
 Usage: oecertdump ENCLAVE_PATH Options
